fix: crash on some operators, location centring, sheet and marker fixes
The crash was a decoding error, not a route problem. bustimes computes
delay and some operators' feeds render it as a float ("delay": -42.0),
but the field was declared as an integer, so kotlinx.serialization threw
while decoding the moment one of those buses was selected. Six of 12,616
vehicles swept were affected, which is why a 17 crashed and a 6 did not.
It is now a Double, with a test that decodes a recorded float-delay
payload and a second, model-independent test documenting why an integer
cannot be used.

The map now moves to the user on every launch rather than only on a first
run. The remembered camera is still used for the first frame, because a
fix can take seconds and a grey rectangle for that long reads as a broken
app, but it is a starting picture rather than the destination. An earlier
change had treated a remembered camera as a reason not to centre at all.

The locate button asked for the cached position first and only fell back
to a fresh one, so once a stale fix was in hand it kept returning there —
which looked like going back to wherever the map started. It now shows
the cached position immediately for responsiveness and then moves again
if a fresh fix turns out to be somewhere else, ignoring differences below
about a hundred metres so the camera does not twitch between two fixes
describing the same place.

A stop's collapsed sheet is now tall enough for the next few departures
rather than the stop name alone; peek height depends on which panel is
showing. The selected bus's heading arrow was hidden underneath its own
enlarged body circle, so the arrow scales up with selection to clear it.

Also fixed while verifying: the journey header showed the next stop as a
raw ATCO code, because the vehicle feed reports it that way. The trip's
calling points are already loaded, so the code is resolved against them.

Verified against live data: a bus whose feed reports a float delay opens
without crashing and reads "1 min late"; launching with a remembered
Cambridge camera and the device in Chester ends up in Chester with the
puck centred; panning away and tapping locate returns there; a stop shows
five departures at peek height.

56 tests, all passing.

+++ app/src/main/java/org/pashri/bustimes/data/model/Vehicle.kt
@@ -23,8 +23,15 @@ data class Vehicle(
     val service: VehicleService? = null,
     val vehicle: VehicleDetail? = null,
     val progress: VehicleProgress? = null,
-    /** Seconds behind schedule; negative means early. */
-    val delay: Int? = null,
+    /**
+     * Seconds behind schedule; negative means early.
+     *
+     * Typed as a Double because it is a computed value and some operators'
+     * feeds yield a fractional one (`-42.0`). Declaring it as an integer made
+     * selecting a bus from those operators throw while decoding, which looked
+     * like a crash on particular routes.
+     */
+    val delay: Double? = null,
     val seats: String? = null,
     val wheelchair: String? = null,
 ) {
